backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
pydantic-settings
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pdfplumber
fuzzywuzzy
python-Levenshtein
sentence-transformers
ollama
edge-tts
aiofiles
pydantic[email]
httpx
